+++ bin/common.inc.bash
@@ -2,5 +2,5 @@
 
 set -eu
 
-readonly DOCKER_CI_IMAGE_NAME="steevanb/php-version-compare:ci"
-readonly DOCKER_RELEASE_IMAGE_NAME="steevanb/php-version-compare:release"
+readonly DOCKER_CI_IMAGE_NAME="steevanb/php-version-comparator:ci"
+readonly DOCKER_RELEASE_IMAGE_NAME="steevanb/php-version-comparator:release"