+++ src/Exception/PhpVersionCompareException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Steevanb\PhpVersionCompare\Exception;
-
-class PhpVersionCompareException extends \Exception
-{
-}