+++ tests/Php72/Comparator/PhpVersionComparator/AssertPhp8Test.php
@@ -1,0 +1,44 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\VersionComparator\Tests\Php72\Comparator\PhpVersionComparator;
+
+use PHPUnit\Framework\TestCase;
+use Steevanb\VersionComparator\{
+    Exception\VersionIsNotBetweenException,
+    Comparator\PhpVersionComparator
+};
+
+final class AssertPhp8Test extends TestCase
+{
+    public function testAssertIsPhp8(): void
+    {
+        $this->expectException(VersionIsNotBetweenException::class);
+        $this->expectExceptionCode(0);
+        $this->expectExceptionMessage(
+            'Version "7.2.34-26+ubuntu20.04.1+deb.sury.org+1" is not between "8.0.0" and "9.0.0".'
+        );
+        PhpVersionComparator::assertIsPhp8();
+    }
+
+    public function testAssertIsPhp80(): void
+    {
+        $this->expectException(VersionIsNotBetweenException::class);
+        $this->expectExceptionCode(0);
+        $this->expectExceptionMessage(
+            'Version "7.2.34-26+ubuntu20.04.1+deb.sury.org+1" is not between "8.0.0" and "8.1.0".'
+        );
+        PhpVersionComparator::assertIsPhp80();
+    }
+
+    public function testAssertIsPhp81(): void
+    {
+        $this->expectException(VersionIsNotBetweenException::class);
+        $this->expectExceptionCode(0);
+        $this->expectExceptionMessage(
+            'Version "7.2.34-26+ubuntu20.04.1+deb.sury.org+1" is not between "8.1.0" and "8.2.0".'
+        );
+        PhpVersionComparator::assertIsPhp81();
+    }
+}