+++ changelog.md
@@ -1,5 +1,8 @@
 ## master
 
+- Rename namespace `Steevanb\PhpVersionCompare\PhpVersion` to `Steevanb\VersionComparator\Comparator`
+- Rename project to `steevanb/version-comparator`
+
 ### [0.1.0](../../compare/0.0.0...0.1.0) - 2021-10-23
 
 - Rename `PhpVersionCompare` to `PhpVersionComparator` and move to `Steevanb\PhpVersionCompare\PhpVersion`